Gold/silverface random item fix ## About The Pull Request
Fixes gold/silverfaces giving random items. What was happening here was that, for an amount of times equal to the number of items in the list of a purchased item, it would pick a random item in that list and drop it on the ground. This is fine for homogenous lists but anything with unique items (old polishing kits we don't use, tea sets, etc etc) had a pretty good chance of not giving you what you wanted. This fixes that.

## Testing Evidence

Fixed this twice on two different servers. Pic is here.

## Why It's Good For The Game

Nobody wants to purchase the fuck-you-death katana for 2.3e+293 mammon and then get two sheathes.